requirements.txt
fastapi=0.109.0
python-dotenv==1.0.1
uvicorn==0.27.0